+++ cli/src/commands/undo.rs
@@ -30,6 +30,8 @@ use crate::ui::Ui;
 /// Undo the last operation
 ///
 /// This undoes the last operation by applying its inverse as a new operation.
+///
+/// There is also a matching `jj redo` command.
 #[derive(clap::Args, Clone, Debug)]
 pub struct UndoArgs {
     /// (deprecated, use `jj op revert <operation>`)
@@ -48,7 +50,7 @@ pub struct UndoArgs {
     what: Vec<RevertWhatToRestore>,
 }
 
-const UNDO_OP_DESC_PREFIX: &str = "undo operation ";
+pub(crate) const UNDO_OP_DESC_PREFIX: &str = "undo operation ";
 
 fn tx_description(op: &Operation) -> String {
     format!("{UNDO_OP_DESC_PREFIX}{}", op.id().hex())
@@ -71,7 +73,9 @@ pub fn cmd_undo(ui: &mut Ui, command: &CommandHelper, args: &UndoArgs) -> Result
 
     let mut op_to_undo = workspace_command.resolve_single_op(&args.operation)?;
 
-    // Growing the "undo-stack" works like this:
+    // Growing the "undo-stack" works as follows. See also the
+    // [redo-stack](./redo.rs), which works in a similar way.
+    //
     // - If the operation to undo is a regular one (not an undo-operation), simply
     //   undo it.
     // - If the operation to undo is an undo-operation itself, try to undo the